moving away from jbuilder for study search results
