--- app/views/api/v1/search/index.json.jbuilder
@@ -1,20 +1,0 @@
-json.set! :type, params[:type]
-json.set! :terms, params[:terms]
-json.set! :term_list, @term_list
-json.set! :current_page, @results.current_page.to_i
-json.set! :total_studies, @results.total_entries
-json.set! :total_pages, @results.total_pages
-json.set! :matching_accessions, @matching_accessions
-json.set! :preset_search, params[:preset_search]
-if @selected_branding_group.present?
-  json.set! :scpbr, @selected_branding_group.name_as_id
-end
-json.facets do
-  json.array! @facets do |facet|
-    json.set! :id, facet[:id]
-    json.set! :filters, facet[:filters]
-  end
-end
-json.studies do
-  json.array! @results, partial: 'api/v1/search/study', as: :study
-end
